Don't try to send an empty batch
When we @report_queue.close in at_exit we generate an "empty" batch, that ends up being []. This means send_report_batch will get called, and make a connection even though there are no reports to send, and bypassing the enabled/disabled logic in the receivers.

If we check for an empty batch we can avoid this connection.