Move enable/disable code to later in the process
Returning early due to the agent being disabled is more efficient, we're not running some code, but it also means that less code is being run locally, or in disabled environments, than would run in an enabled environment, say production. By moving the disabled check later we run more of the agent code and this will help highlight any problems or incompatibilities in with the applicatino code earlier, in more environments where the agent is disabled, such as dev.