Stop Warning injection being magical
Previously when the lib/control/ruby_reporting file was loaded it triggered injection of the RubyReporting module into Warning's ancestors. This is a bit magical, and hard to test.

Lets hoist the injection logic into the `insert_ruby_deprecation_behaviour` method, and also make it easy to inject an object to insert into for testing purposes