Hook into Rails >= 7.1 without causing warnings
All warnings going through the old singleton API are forwarded through the new deprecators instance(s), so we don't need to hook both once the new APIs are in play. Switch on the existence of that within a Rails app. Annoyingly there's no easy way to "append" a new behavior to the collection, only replace it entirely so we've accessed a private instance variable to keep any existing setting.

This will cause issues when we go non-rails, but that's a future problem.